Revert "Dev/profile-image  프로필이미지 컴포넌트 함수분리"
