+++ src/components/loading/Loading.css
@@ -1,6 +1,3 @@
-body {
-  max-width: none;
-}
 .loading-container {
   width: 100vw;
   height: 100vh;
