+++ src/pages/userinfo/UserInfo.js
@@ -1,16 +1,36 @@
 import './UserInfo.css';
 import ProfileImage from '../../components/profileImage/ProfileImage';
+import { Validator } from './Validator';
+
+import { UserInfoDTO } from './userInfoDTO';
+import { EmployeeListFetch } from '../employeeListTable/EmployeeListFetch';
 import defalutProfileImg from '../../assets/images/avatar-default.jpg';
+
 import { EmployeeListTable } from '../employeeListTable/EmployeeListTable';
 import Mypage from '../mypage/Mypage';
-import { checkIsFormAllInputsValid, setSaveButtonElemDisable } from './validatorHelpers';
-import { setupButtonHandlers, handleFocusOutEventForValidCheck } from './eventHandlers';
-import { saveUserDataOnSessionStorage, saveUserDataWhenFormSubmit } from './dataManagement';
-import { setFormIdInputDisable, toggleFormMode, showDelaySubmitMessage } from './uiHelpers';
-import { loadUserFromSession, loadUserFromProps } from './dataLoaders';
+
+const validator = new Validator();
+
+const isValidAndReturnMessage = {
+  'user-id': (inputValue) => validator.idValidator(inputValue),
+  'user-duplicate-id': (inputValue) => validator.idDuplicateValidator(inputValue),
+  'user-email': (inputValue) => validator.emailValidator(inputValue),
+  'user-password': (inputValue) => validator.passwordValidator(inputValue),
+  'user-phone': (inputValue) => validator.phoneValidator(inputValue),
+};
+
+const isValidInputState = {
+  'user-id': false,
+  'user-duplicate-id': false,
+  'user-email': false,
+  'user-password': false,
+  'user-phone': false,
+};
 
 export default class UserInfo {
   constructor(cotainer, props = {}) {
+    this.employeeListFetch = new EmployeeListFetch();
+    this.validator = new Validator();
     this.btnState = false;
 
     const {
@@ -44,7 +64,14 @@ export default class UserInfo {
         },
       },
     };
-
+    this.fields = [
+      { id: 'user-id', key: 'id' },
+      { id: 'user-password', key: 'userPassword' },
+      { id: 'user-name', key: 'name' },
+      { id: 'user-email', key: 'email' },
+      { id: 'user-phone', key: 'phone' },
+      { id: 'user-position', key: 'position' },
+    ];
     this.el = cotainer;
   }
   render() {
@@ -172,11 +199,11 @@ export default class UserInfo {
 
     // 수정, 조회 페이지 전환
     // 수정 모드와 조회 모드 전환
-    toggleFormMode(this.el, this.info);
+    this.toggleFormMode();
 
     // 버튼 핸들러 설정 적용
     //  수정, 취소 버튼 클릭 이벤트 헨들러
-    setupButtonHandlers(this.el, this.switchToEditMode.bind(this), history.back.bind(history));
+    this.setupButtonHandlers();
 
     // 프로필 이미지 컴포넌트 불러오기
     // 직원에 등록된 프로필이미지 적용하기
@@ -185,10 +212,18 @@ export default class UserInfo {
 
   // 유저모드, 관리자 모드 전환
   setAdminModeOrUserMode() {
+    // 관리자 모드
+    if (this.isAdminPage()) {
+      // 아이디 입력 활성화
+      // 아이디 중복성 체크 활성화
+      this.btnType();
+      return;
+    }
+
     // 유저모드
     if (!this.isAdminPage()) {
       // 아이디 입력 비활성화
-      setFormIdInputDisable(this.el);
+      this.setFormIdInputDisable();
       return;
     }
   }
@@ -199,18 +234,27 @@ export default class UserInfo {
     // submit이 가능하게함.
     if (this.isEditOrRegisterPage()) {
       // 저장 버튼 클릭 비활성화
-      setSaveButtonElemDisable();
+      this.setSaveButtonElemDisable();
 
       // focusout 이벤트를 헨들러를 붙임.
-      handleFocusOutEventForValidCheck();
+      this.handleFocusOutEventForValidCheck();
 
       // 현재 화면에서 유효성을 체크함.
-      checkIsFormAllInputsValid(this.getFormInputData, this.isAdminPage, this.isEditOrRegisterPage);
+      this.checkIsFormAllInputsValid();
 
       // 폼 submit 이벤트 헨들러
       // form Submit 이벤트를 이용해서 백엔드에 직원 정보 수정, 등록 데이터를 전송함.
       this.handleFormSubmitEvent();
     }
+  }
+
+  // ID 입력 차단
+  setFormIdInputDisable() {
+    const userId = this.el.querySelector('#user-id');
+    userId.readOnly = true;
+    userId.style.border = 'none';
+    const duplicateValidedButton = document.querySelector('.user-info__type');
+    duplicateValidedButton.style.display = 'none';
   }
 
   // 관리자 페이지인지 아닌지를 체크함.
@@ -223,6 +267,13 @@ export default class UserInfo {
     return this.info === '수정' || this.info === '등록';
   }
 
+  // 유저가 저장을 클릭하고 나서 백엔드에 데이터가 전송 되기 까지 500ms 딜레이가 있을 때
+  // 저장 버튼 아래 "임직원 변경 사항 저장중.." 메시지가 뜨게함.
+  showDelaySubmitMessage() {
+    const delay_submit_message = document.querySelector('.user-info-submit-delay__message');
+    delay_submit_message.classList.add('toggle-f');
+  }
+
   // form Submit 이벤트 리스너
   // form 데이터를 백엔드에 전송함.
   onFormSubmit = async (event) => {
@@ -230,19 +281,16 @@ export default class UserInfo {
     const form = this.el.querySelector('.user-info');
     // 만약 폼에 입력된 값이 하나라도 유효성 검사 실패했다면
     //  submit 이벤트를 중지시킴.
-    if (
-      !checkIsFormAllInputsValid(this.getFormInputData, this.isAdminPage, this.isEditOrRegisterPage)
-    )
-      return;
+    if (!this.checkIsFormAllInputsValid) return;
 
     // `저장 버튼` 아래 "임직원 변경 사항 저장중.." 메시지가 뜨게함.
-    showDelaySubmitMessage(this.el);
+    this.showDelaySubmitMessage();
 
     // SessionStorage에 변경된 직원 데이터(직원 아이디,  비밀번호, 이름, 이메일, 직급)을 저장함.
-    saveUserDataOnSessionStorage(new FormData(form));
+    this.saveUserDataOnSessionStorage(form);
 
     // 변경된 직원 데이터를 DB에 저장.
-    await saveUserDataWhenFormSubmit(form, this.info);
+    await this.saveUserDataWhenFormSubmit(form);
 
     // submit 전송하고 db에 반영되기 까지 0.3초 시간을 기다린다음.
     // 유저(직원)일 경우 마이페이지로 이동하고,
@@ -273,6 +321,16 @@ export default class UserInfo {
     if (e.target.classList.contains('user-info__btn--cancel')) return history.back();
   };
 
+  // 버튼 헨들러
+  // 수정, 취소 버튼 클릭 이벤트 헨들러
+  setupButtonHandlers() {
+    // editAndCancelbuttons는 수정, 취소 버튼들임.
+    const editAndCancelbuttons = this.el.querySelectorAll('.user-info__btns > button');
+    editAndCancelbuttons.forEach((btn) =>
+      btn.addEventListener('click', this.onClickEditButtonOrShowButton),
+    );
+  }
+
   // 다른 페이지로 이동
   renderEmployeeListPageOrMyPage() {
     // 관리자일 경우 /employee-list 직원 관리페이지로 이동
@@ -292,14 +350,84 @@ export default class UserInfo {
     }
   }
 
+  // submit 헨들러
+  //  백엔드로 유저 데이터 수정 요청 API 호출
+  //  백엔드로 유저 데이터 등록 요청 API 호출
+  async saveUserDataWhenFormSubmit(form) {
+    const formData = new FormData(form);
+    const userInfotr = Object.fromEntries(formData.entries());
+    const trdataId = document.querySelector('.user-info__lists-wrap');
+    const employeeListFetch = new EmployeeListFetch();
+    userInfotr['data-id'] = trdataId.dataset.dataId; // 데이터 아이디 설정
+    userInfotr['data-profileImg'] = document.getElementById('img1').dataset.profileImage;
+
+    if (this.info === '수정') {
+      await employeeListFetch.updateEmployee(new UserInfoDTO(userInfotr));
+    } else if (this.info === '등록') {
+      await employeeListFetch.addEmployee(new UserInfoDTO(userInfotr));
+    }
+  }
   // 유저정보 수정
 
+  // 유저 정보를 SessionStorage에 저장
+  saveUserDataOnSessionStorage(props) {
+    const formData = new FormData(props); // 폼 데이터 가져오기
+    this.fields.forEach(({ id, key }) => {
+      const value = formData.get(id); // 폼 데이터에서 값 가져오기
+      if (value !== null) {
+        sessionStorage.setItem(key, value);
+      }
+    });
+  }
+
+  // 수정 모드와 조회 모드 전환
+  toggleFormMode() {
+    const sel = this.el.querySelector('#user-position');
+    const [editInfoButton, saveCancelButton] = this.el.querySelectorAll('.user-info__btns');
+    if (this.info === '조회') {
+      const inputEl = this.el.querySelectorAll('input');
+      this.el.querySelector('.user-info__type').classList.add('btnNone');
+      inputEl.forEach((el) => {
+        el.readOnly = true;
+        el.style.border = 'none';
+        sel.disabled = true;
+        editInfoButton.style.display = 'none';
+      });
+    } else {
+      saveCancelButton.style.display = 'none';
+    }
+  }
   // 유저 정보 불러오기
   loadUserData() {
     if (this.permission === 'user') {
-      loadUserFromSession(this.el);
+      this.loadUserFromSession(this.fields);
     } else {
-      loadUserFromProps(this.el, this.state, this.userid);
+      this.loadUserFromProps(this.fields);
+    }
+  }
+
+  // user일경우 세션스토리지에서 정보 가져오기
+  loadUserFromSession() {
+    sessionStorage.setItem('userPassword', 'password');
+    this.state = { ...sessionStorage };
+    if (this.state) {
+      this.fields.forEach(({ id, key }) => {
+        this.el.querySelector(`#${id}`).value = this.state[key] || '';
+      });
+    }
+  }
+
+  // admin일경우 props에서 정보 가져오기
+  loadUserFromProps() {
+    if (this.state.user && this.state.user[this.userid]) {
+      const trdataId = document.querySelector('.user-info__lists-wrap');
+      if (this.state.user[this.userid] === undefined) {
+        console.error('dataId가 없습니다.');
+      }
+      trdataId.dataset.dataId = this.state.user[this.userid].dataId || '';
+      this.fields.forEach(({ id, key }) => {
+        this.el.querySelector(`#${id}`).value = this.state.user[this.userid][key] || '';
+      });
     }
   }
 
@@ -320,8 +448,78 @@ export default class UserInfo {
     };
   };
 
+  // 처음 유저가 수정, 등록창에 들어오면, 저장 버튼을 못누르게 함
+  setSaveButtonElemDisable = () => {
+    const saveButtonElem = document.querySelector('.user-info__btn--save');
+    saveButtonElem.disabled = true;
+    saveButtonElem.classList.add('user-info__btn--disable');
+  };
+
+  // 유저가 폼에 입력을 제대로 입력했으면, 저장 버튼을 누를 수 있게함.
+  setSaveButtonElemAble = () => {
+    const saveButtonElem = document.querySelector('.user-info__btn--save');
+    saveButtonElem.disabled = false;
+    saveButtonElem.classList.remove('user-info__btn--disable');
+  };
+
+  // 유저가 form input에 대한 각각 검사를 함.
+  // inputElemId: String, input 요소 id 이름
+  // inputElemValue: String, input.value 값
+  validCheckerFormInput = (inputElemId, inputElemValue) => {
+    const returnValidMessageFunc = isValidAndReturnMessage[inputElemId];
+    const validMessage = returnValidMessageFunc(inputElemValue);
+
+    const errorcheck = document.querySelector(`.${inputElemId}__error`);
+    errorcheck.innerHTML = returnValidMessageFunc(inputElemValue);
+
+    if (validMessage === 'success') {
+      errorcheck.classList.remove('-fail-red');
+      errorcheck.classList.add('-success-green');
+      isValidInputState[inputElemId] = true;
+    }
+    if (validMessage !== 'success') {
+      errorcheck.classList.remove('-success-green');
+      errorcheck.classList.add('-fail-red');
+      isValidInputState[inputElemId] = false;
+    }
+
+    // 유저가 폼에 입력을 모두 제대로 입력했으면, 저장 버튼을 누를 수 있게함.
+    if (this.isValidStateAllTrue()) {
+      this.setSaveButtonElemAble();
+    }
+  };
+
+  // 모든 유효성 검증에 대한 상태값들이 모두 true인지 확인.
+  isValidStateAllTrue() {
+    return Object.values(isValidInputState).every(Boolean);
+  }
+
+  //  focusOut 발생시 마다 제대로 입력 되었는지 체크를함
+  // 아이디, 패스워드, 이메일, 핸드폰 번호에 대한 각각의 유효성 검사를함.
+  // 전체 이들 모두가 유효성 검사를 통과 해야 저장 버튼이 활성화됨
+  checkIsFormAllInputsValid = () => {
+    const formInputDatas = this.getFormInputData();
+    this.isAdminPage() && this.isEditOrRegisterPage()
+      ? this.validCheckerFormInput('user-id', formInputDatas['user-id'])
+      : (isValidInputState['user-id'] = true);
+    this.validCheckerFormInput('user-password', formInputDatas['user-password']);
+    this.validCheckerFormInput('user-email', formInputDatas['user-email']);
+    this.validCheckerFormInput('user-phone', formInputDatas['user-phone']);
+    return this.isValidStateAllTrue();
+  };
+
+  // focusOut 이벤트헨들러,  발생시 마다 제대로 입력 되었는지 체크를함
+  // onFocusOutEvent 함수를 form에 붙임.
+  handleFocusOutEventForValidCheck = () => {
+    const userformElem = document.querySelector('.user-info');
+    userformElem.addEventListener('focusout', this.checkIsFormAllInputsValid);
+  };
+
   // ID 중복 확인 함수
   async validatorIdDuplicate(employeeId) {
     return await this.employeeListFetch.getEmployeeListById(employeeId);
   }
+
+  // 아이디 중복 확인 버튼 로직, 이벤트 헨들러
+  btnType() {}
 }
