--- src/pages/userinfo/constants.js
@@ -1,10 +1,0 @@
-// 클래스 내부 상수 분리
-
-export const fields = [
-  { id: 'user-id', key: 'id' },
-  { id: 'user-password', key: 'userPassword' },
-  { id: 'user-name', key: 'name' },
-  { id: 'user-email', key: 'email' },
-  { id: 'user-phone', key: 'phone' },
-  { id: 'user-position', key: 'position' },
-];
